fix fixes :D and more
