+++ src/components/Visualizations/Table/HeroImage.css
@@ -1,8 +1,4 @@
-@import "../palette.css";
-
-:export {
-  blue:   var(--primaryLinkColor);
-}
+@import "../../palette.css";
 
 .actionDoneAll {
   position: relative !important;
@@ -13,6 +9,7 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
+  line-height: 1.2;
 }
 
 .image {
