+++ src/components/Visualizations/Table/KDA.css
@@ -1,11 +1,11 @@
-@import '../palette.css';
+@import '../../palette.css';
 
 .container {
   position: relative;
 }
 
 .percent {
-  composes: TablePercent from "./TablePercent.css";
+  composes: TablePercent from "./Percent.css";
   width: calc(300% + 36px);
   left: -7px;
 }
