+++ src/components/Visualizations/Table/Percent.css
