+++ src/components/Visualizations/index.js
@@ -1,4 +1,5 @@
-export { default as TablePercent } from './TablePercent';
-export { default as KDA } from './KDA';
+export { default as TablePercent } from './Table/Percent';
+export { default as KDA } from './Table/KDA';
 export { default as Graph } from './Graph';
-export { default as TableHeroImage } from './TableHeroImage';
+export { default as TableHeroImage } from './Table/HeroImage';
+export { default as FromNowTooltip } from './FromNowTooltip';
