+++ src/lang/en.json
@@ -15,6 +15,7 @@
   "abbr_hero_damage_per_min": "HDM",
   "abbr_hero_healing_per_min": "HHM",
   "abbr_tower_damage_per_min": "TDM",
+  "abbr_kda": "KDA",
   "actions_per_min": "Actions performed by player per minute",
   "actions_per_min_abbr": "APM",
   "rune_0": "Double Damage",
@@ -98,6 +99,7 @@
   "purchase_buyback": "Buyback purchases",
   "used_ward_observer": "Number of Observer Wards placed during the game",
   "used_ward_sentry": "Number of Sentry Wards placed during the game",
+  "parsed": "Replay has been parsed for additional statistics",
   "hero_id": "The hero played by player",
   "result": "Whether the player won or lost",
   "match_id": "The ID of the match",
